% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{calc_expressed_genes}
\alias{calc_expressed_genes}
\title{Check}
\usage{
calc_expressed_genes(sce, params)
}
\arguments{
\item{sce}{SingleCellExperiment class containing all cells and genes required}

\item{params}{List of all parameters specified.}
}
\description{
Check
}
\keyword{internal}
