% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{check_params}
\alias{check_params}
\title{check all parameters}
\usage{
check_params(x, y, y_labels, assay_type, sel_genes, gene_list, scale,
  smooth, min_expression, penalization, method, score, n_folds, test_propn,
  lambdas, max_iters, seed)
}
\value{
list of validated parameters
}
\description{
check all parameters
}
\keyword{internal}
