% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{do_topgo_for_cluster}
\alias{do_topgo_for_cluster}
\title{Calculate GO enrichment for each cluster vs all other genes}
\usage{
do_topgo_for_cluster(clusters_dt, sig_cutoff, org_mapping)
}
\arguments{
\item{clusters_dt}{}

\item{sig_cutoff}{How many genes should be in the cluster for us to consider a GO term?}
}
\value{
data.table with GO term results
}
\description{
Calculate GO enrichment for each cluster vs all other genes
}
