% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{.calc_clusters_dt}
\alias{.calc_clusters_dt}
\title{make nice data.table of hierarchical clusters}
\usage{
.calc_clusters_dt(hclust_obj, x_data, proj_dt, k = 5)
}
\arguments{
\item{hclust_obj}{Result of hclust}

\item{x_data}{Data used to calculate psuper_obj}

\item{proj_dt}{Projection of cells onto psupertime}
}
\value{
data.table containing clusters of genes, ordered according to correlation with psupertime
}
\description{
make nice data.table of hierarchical clusters
}
