% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{.calc_expressed_genes}
\alias{.calc_expressed_genes}
\title{Restrict to genes with minimum proportion of expression defined in ps_params$min_expression}
\usage{
.calc_expressed_genes(x, ps_params)
}
\arguments{
\item{x}{SingleCellExperiment class containing all cells and genes required}

\item{ps_params}{List of all parameters specified.}
}
\description{
Restrict to genes with minimum proportion of expression defined in ps_params$min_expression
}
\keyword{internal}
