% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{.calc_hvg_genes}
\alias{.calc_hvg_genes}
\title{Calculates list of highly variable genes (according to approach in scran).}
\usage{
.calc_hvg_genes(sce, ps_params, do_plot = FALSE)
}
\arguments{
\item{ps_params}{List of all parameters specified.}

\item{x}{SingleCellExperiment class or matrix of log counts}
}
\description{
Calculates list of highly variable genes (according to approach in scran).
}
\keyword{internal}
