% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{.check_conf_params}
\alias{.check_conf_params}
\title{Check variables for confusion matrices}
\usage{
.check_conf_params(plot_var)
}
\arguments{
\item{plot_var}{Variable to plot: prop_true is proportion of true labels, prop_predict is proportion of predicted labels, N is # of cells}
}
\value{
list with checked plot_var, and nice label
}
\description{
Check variables for confusion matrices
}
