% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{.get_test_idx}
\alias{.get_test_idx}
\title{Get list of cells to keep aside as test set}
\usage{
.get_test_idx(y, ps_params)
}
\arguments{
\item{y}{list of y labels}
}
\value{
Indices for test set
}
\description{
Get list of cells to keep aside as test set
}
\keyword{internal}
