% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{.get_tf_list}
\alias{.get_tf_list}
\title{Get list of transcription factors}
\usage{
.get_tf_list(dirs)
}
\value{
List of all transcription factors specified.
}
\description{
Get list of transcription factors
}
\keyword{internal}
