% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{.glmnetcr_propn}
\alias{.glmnetcr_propn}
\title{This is based on an equivalent function from the package \code{glmnetcr}, 
which is sadly no longer on CRAN.}
\usage{
.glmnetcr_propn(
  x,
  y,
  method = "proportional",
  weights = NULL,
  offset = NULL,
  alpha = 1,
  nlambda = 100,
  lambda.min.ratio = NULL,
  lambda = NULL,
  standardize = TRUE,
  thresh = 1e-04,
  exclude = NULL,
  penalty.factor = NULL,
  maxit = 100
)
}
\description{
This is based on an equivalent function from the package \code{glmnetcr}, 
which is sadly no longer on CRAN.
}
\keyword{internal}
