% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{.make_col_vals}
\alias{.make_col_vals}
\title{Define RColorBrewer palette to use; default is RdBu.}
\usage{
.make_col_vals(y_labels, palette = "RdBu")
}
\arguments{
\item{y_labels}{List of labels used for training}
}
\value{
Colour values
}
\description{
Define RColorBrewer palette to use; default is RdBu.
}
\keyword{internal}
