% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{.make_plot_dt}
\alias{.make_plot_dt}
\title{Internal function}
\usage{
.make_plot_dt(x_data, hclust_obj, proj_dt, clusters_dt)
}
\arguments{
\item{x_data}{}

\item{hclust_obj}{}

\item{proj_dt}{}

\item{clusters_dt}{}
}
\value{
data.table for plotting
}
\description{
Internal function
}
