% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{.make_x_data}
\alias{.make_x_data}
\title{Process input data}
\usage{
.make_x_data(x, sel_genes, ps_params)
}
\arguments{
\item{x}{SingleCellExperiment or matrix of log counts}

\item{sel_genes}{Selected genes}

\item{ps_params}{Full list of parameters}
}
\value{
Matrix of dimension # cells by # selected genes
}
\description{
Note that input is matrix with rows=genes, cols=cells, and that output
has rows=cells, genes=cols
}
\keyword{internal}
