% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{.restrict_to_y_labels}
\alias{.restrict_to_y_labels}
\title{Use y_labels to define cells to use, and order of labels}
\usage{
.restrict_to_y_labels(x_data, y, y_labels)
}
\arguments{
\item{x_data}{matrix output from make_x_data (rows=cells, cols=genes)}

\item{y}{factor of cell labels}

\item{y_labels}{list of labels to restrict to, and order to use}
}
\description{
Use y_labels to define cells to use, and order of labels
}
