% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{.select_genes}
\alias{.select_genes}
\title{Select genes for use in regression}
\usage{
.select_genes(x, ps_params)
}
\arguments{
\item{x}{SingleCellExperiment class containing all cells and genes required, or matrix of counts}

\item{ps_params}{List of all parameters specified.}
}
\description{
Select genes for use in regression
}
\keyword{internal}
