% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{double_psupertime}
\alias{double_psupertime}
\title{Projects two different psupertimes onto each other}
\usage{
double_psupertime(psuper_1, psuper_2, run_names = NULL, process = FALSE)
}
\arguments{
\item{psuper_1, }{psuper_2 Two previously calculated psupertime objects}

\item{labels}{Character vector of length two, labelling the psupertime inputs}
}
\value{
data.table containing projections in both directions
}
\description{
Projects two different psupertimes onto each other
}
