% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{make_best_beta}
\alias{make_best_beta}
\title{Extracts best coefficients.}
\usage{
make_best_beta(glmnet_best, best_lambdas)
}
\value{
data.table containing learned coefficients for all genes used as input.
}
\description{
Extracts best coefficients.
}
\keyword{internal}
