% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_double_psupertime_contour}
\alias{plot_double_psupertime_contour}
\title{Projects two different psupertimes on top of each other}
\usage{
plot_double_psupertime_contour(
  double_obj = NULL,
  psuper_1 = NULL,
  psuper_2 = NULL,
  run_names = NULL
)
}
\arguments{
\item{double_obj}{Result of applying double_psupertime to two previously calculated psupertime objects}

\item{psuper_1, }{psuper_2 Two previously calculated psupertime objects}

\item{run_names}{Character vector of length two, labelling the psupertime inputs}
}
\value{
ggplot object plotting the two against each other
}
\description{
See `plot_double_psupertime` for further detail.
}
