% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_double_psupertime_genes}
\alias{plot_double_psupertime_genes}
\title{Compares coefficients for genes learned from different psupertimes}
\usage{
plot_double_psupertime_genes(psuper_1, psuper_2, run_names = NULL)
}
\arguments{
\item{psuper_1, }{psuper_2 Two previously calculated psupertime objects}

\item{run_names}{Character vector of length two, labelling the psupertime inputs}
}
\value{
ggplot object plotting the two sets of coefficients
}
\description{
Compares coefficients for genes learned from different psupertimes
}
