% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_go_results}
\alias{plot_go_results}
\title{Plots the significant GO terms for each cluster}
\usage{
plot_go_results(go_list, sig_cutoff = 5, p_cutoff = 0.1)
}
\arguments{
\item{sig_cutoff}{What is the minimum number of annotated genes to display a GO term?}

\item{p_cutoff}{What is the maximum p-value to display a GO term?}

\item{go_results}{Output from GO analysis}
}
\value{
bar plot
}
\description{
Plots the significant GO terms for each cluster
}
