% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_heatmap_of_gene_clusters}
\alias{plot_heatmap_of_gene_clusters}
\title{Plot heatmap of gene clusters}
\usage{
plot_heatmap_of_gene_clusters(go_list)
}
\arguments{
\item{go_list}{Output from GO analysis}
}
\value{
ggplot object
}
\description{
Plot heatmap of gene clusters
}
