% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_identified_gene_coefficients}
\alias{plot_identified_gene_coefficients}
\title{Plots top coefficients}
\usage{
plot_identified_gene_coefficients(psuper_obj, abs_cutoff = 0.1)
}
\arguments{
\item{psuper_obj}{Psupertime object, output from psupertime}
}
\value{
ggplot2 object
}
\description{
Plots top coefficients
}
