% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_identified_genes_over_psupertime}
\alias{plot_identified_genes_over_psupertime}
\title{Plots profiles of identified genes against psupertime.}
\usage{
plot_identified_genes_over_psupertime(psuper_obj, n_to_plot = 25,
  palette = "RdBu")
}
\arguments{
\item{psuper_obj}{Psupertime object, output from psupertime}
}
\value{
ggplot2 object
}
\description{
Plots profiles of identified genes against psupertime.
}
