% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_profiles_of_gene_clusters}
\alias{plot_profiles_of_gene_clusters}
\title{Plot heatmap of gene clusters}
\usage{
plot_profiles_of_gene_clusters(go_list, label_name = "Ordered labels",
  palette = "RdBu")
}
\arguments{
\item{go_list}{Output from GO analysis}

\item{label_name}{Description for the ordered labels in the legend (e.g. 'Developmental stage (days)')}

\item{palette}{RColorBrewer palette to use}
}
\value{
ggplot object
}
\description{
Plot heatmap of gene clusters
}
