% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_specified_genes_over_psupertime}
\alias{plot_specified_genes_over_psupertime}
\title{Plots profiles of hand-selected genes against psupertime.}
\usage{
plot_specified_genes_over_psupertime(psuper_obj, extra_genes,
  palette = "RdBu")
}
\arguments{
\item{psuper_obj}{Psupertime object, output from psupertime}
}
\value{
ggplot2 object
}
\description{
Plots profiles of hand-selected genes against psupertime.
}
