% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_train_results}
\alias{plot_train_results}
\title{Plot results of training}
\usage{
plot_train_results(psuper_obj)
}
\arguments{
\item{psuper_obj}{Psupertime object, output from psupertime}
}
\value{
ggplot2 object showing test and training performance of classifier.
}
\description{
Plot results of training
}
