% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{project_onto_psupertime}
\alias{project_onto_psupertime}
\title{Gives projection of data onto psupertime (either using original data, or new data)}
\usage{
project_onto_psupertime(
  psuper_obj,
  new_x = NULL,
  new_y = NULL,
  process = FALSE
)
}
\arguments{
\item{psuper_obj}{Psupertime object, output from psupertime}

\item{new_x, }{new_y Optional pair of new data and labels}
}
\value{
data.table with projection and labels
}
\description{
Gives projection of data onto psupertime (either using original data, or new data)
}
