% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tf_human}
\alias{tf_human}
\title{List of human transcription factors}
\format{
A character vector of length 795
}
\source{
\url{https://www.grnpedia.org/trrust/downloadnetwork.php}
}
\usage{
tf_human
}
\description{
Derived from the TRRUST project: https://www.grnpedia.org/trrust/
}
\keyword{datasets}
