% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tf_list}
\alias{tf_list}
\title{List of transcription factors}
\format{A character vector of length 853}
\source{
\url{}
}
\usage{
tf_list
}
\description{
Derived from symbols annotated with GO term 'GO:0003700' in ontology molecular function
}
\keyword{datasets}
