% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tf_mouse}
\alias{tf_mouse}
\title{List of mouse transcription factors}
\format{A character vector of length 827}
\source{
\url{https://www.grnpedia.org/trrust/downloadnetwork.php}
}
\usage{
tf_mouse
}
\description{
Derived from the TRRUST project: https://www.grnpedia.org/trrust/
}
\keyword{datasets}
