% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_identified_genes_over_psupertime}
\alias{plot_identified_genes_over_psupertime}
\title{Plots profiles of identified genes against psupertime.}
\usage{
plot_identified_genes_over_psupertime(psuper_obj,
  label_name = "Ordered labels", n_to_plot = 20, palette = "RdBu")
}
\arguments{
\item{psuper_obj}{Psupertime object, output from psupertime}

\item{label_name}{Description for the ordered labels in the legend (e.g. 'Developmental stage (days)')}

\item{n_to_plot}{Maximum number of genes to plot (default 20)}

\item{palette}{RColorBrewer palette to use}
}
\value{
ggplot2 object
}
\description{
Plots profiles of identified genes against psupertime.
}
