% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_labels_over_psupertime}
\alias{plot_labels_over_psupertime}
\title{Plots labels over their projected values on psupertime.}
\usage{
plot_labels_over_psupertime(psuper_obj, palette = "RdBu")
}
\arguments{
\item{psuper_obj}{Psupertime object, output from psupertime}
}
\value{
ggplot2 object
}
\description{
Plots labels over their projected values on psupertime.
}
