% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_new_data_over_psupertime}
\alias{plot_new_data_over_psupertime}
\title{Plots profiles of hand-selected genes against psupertime.}
\usage{
plot_new_data_over_psupertime(psuper_obj, new_x, new_y, palette = "BrBG")
}
\arguments{
\item{psuper_obj}{Psupertime object, output from psupertime}

\item{new_x, new_y}{Optional data to predict from psuper_obj}

\item{plot_var}{Variable to plot: prop_true is proportion of true labels, prop_predict is proportion of predicted labels, N is # of cells}
}
\value{
ggplot2 object
}
\description{
Plots profiles of hand-selected genes against psupertime.
}
