% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{plot_specified_genes_over_psupertime}
\alias{plot_specified_genes_over_psupertime}
\title{Plots profiles of hand-selected genes against psupertime.}
\usage{
plot_specified_genes_over_psupertime(psuper_obj, extra_genes,
  label_name = "Ordered labels", palette = "RdBu")
}
\arguments{
\item{psuper_obj}{psupertime object, output from psupertime}

\item{extra_genes}{List of genes to be plotted (these must be in the set of genes used for calculating psupertime, e.g. highly variable genes)}

\item{label_name}{Description for the ordered labels in the legend (e.g. 'Developmental stage (days)')}

\item{palette}{RColorBrewer palette to use}
}
\value{
ggplot2 object
}
\description{
Plots profiles of hand-selected genes against psupertime.
}
