% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime.R
\name{psupertime}
\alias{psupertime}
\title{Supervised pseudotime}
\usage{
psupertime(x, y, y_labels = NULL, sel_genes = "hvg",
  gene_list = NULL, scale = TRUE, smooth = TRUE,
  min_expression = 0.01, penalization = "1se", method = "cumulative",
  score = "xentropy", n_folds = 5, test_propn = 0.1,
  lambdas = NULL, max_iters = 1000, seed = 1234)
}
\arguments{
\item{x}{Either SingleCellExperiment class containing all cells and genes required, or matrix of log TPM values.}

\item{y}{Vector of labels, which should have same length as number of columns in sce / x. Factor levels will be taken as the intended order for training.}

\item{y_labels}{Alternative ordering of the labels in y. All labels must be present in y. May be a subset of the labels in y.}

\item{sel_genes}{Subset of genes to be used for training.}
}
\value{
psupertime object
}
\description{
Supervised pseudotime
}
