% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{psupertime_go_analysis}
\alias{psupertime_go_analysis}
\title{GO enrichment analysis for genes learned from different psupertimes}
\usage{
psupertime_go_analysis(psuper_obj, org_mapping = "org.Mm.eg.db")
}
\arguments{
\item{psuper_obj}{A previously calculated psupertime object}

\item{org_mapping}{Organism to use for annotations}
}
\value{
data.table containing results of GO enrichment analysis
}
\description{
GO enrichment analysis for genes learned from different psupertimes
}
