% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psupertime_plots.R
\name{psupertime_plot_all}
\alias{psupertime_plot_all}
\title{Convenience function to do multiple plots}
\usage{
psupertime_plot_all(psuper_obj, output_dir = ".", tag = "",
  ext = "png")
}
\arguments{
\item{psuper_obj}{Psupertime object, output from psupertime}

\item{output_dir}{Directory to save to}

\item{tag}{Label for all files}

\item{ext}{Image format for outputs, compatible with ggsave (eps, ps, tex, pdf, jpeg, tiff, png, bmp, svg, wmf)}
}
\description{
Convenience function to do multiple plots
}
